requirements.txt
appnope==0.1.0
cycler==0.9.0
decorator==4.0.6
gnureadline==6.3.3
ipython==4.0.1
ipython-genutils==0.1.0
jieba==0.38
matplotlib==1.5.0
numpy==1.10.4
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.1.0
ptyprocess==0.5
pyparsing==2.0.7
python-dateutil==2.4.2
pytz==2015.7
simplegeneric==0.8.1
six==1.10.0
snownlp==0.12.3
traitlets==4.0.0
wheel==0.24.0
wordcloud==1.2
